Dev: Refactor code to use factory function
- Blueprints available for auth, user, main, errors
