+++ app/__init__.py
@@ -12,59 +12,81 @@ from config import Config
 from sqlalchemy import MetaData
 
 convention = {
-    "ix": 'ix_%(column_0_label)s',
-    "uq": "uq_%(table_name)s_%(column_0_name)s",
-    "ck": "ck_%(table_name)s_%(constraint_name)s",
-    "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
-    "pk": "pk_%(table_name)s"
-}
+        "ix": 'ix_%(column_0_label)s',
+        "uq": "uq_%(table_name)s_%(column_0_name)s",
+        "ck": "ck_%(table_name)s_%(constraint_name)s",
+        "fk": "fk_%(table_name)s_%(column_0_name)s_%(referred_table_name)s",
+        "pk": "pk_%(table_name)s"
+    }
 metadata = MetaData(naming_convention=convention)
 
+db = SQLAlchemy(metadata=metadata)
+migrate = Migrate()
+moment = Moment()
+login = LoginManager()
+login.login_view = 'auth.login'
+mail = Mail()
+boostrap = Bootstrap()
 
-app = Flask(__name__)
-app.config.from_object(Config)
+def create_app(config_class=Config):
+    app = Flask(__name__)
+    app.config.from_object(config_class)   
 
-db = SQLAlchemy(app, metadata=metadata)
-migrate = Migrate(app, db, render_as_batch=True)
-moment = Moment(app)
-login = LoginManager(app)
-login.login_view = 'login'
-mail = Mail(app)
-boostrap = Bootstrap(app)
+    from app.errors import bp as errors_bp
+    app.register_blueprint(errors_bp)
 
-if not app.debug:
-    if app.config['MAIL_SERVER']:
-        auth = None
-        if app.config['MAIL_USERNAME'] or app.config['MAIL_PASSWORD']:
-            auth = (app.config['MAIL_USERNAME'], app.config['MAIL_PASSWORD'])
-        secure = None
-        if app.config['MAIL_USE_TLS']:
-            secure = ()
-        mail_handler = SMTPHandler(
-            mailhost=(app.config['MAIL_SERVER'], app.config['MAIL_PORT']),
-            # fromaddr='noreply@' + app.config['MAIL_SERVER'],
-            fromaddr=app.config['MAIL_DEFAULT_SENDER'],
-            toaddrs=app.config['ADMINS'],
-            subject='Taskmate Failure',
-            credentials=auth, secure=secure)
-        mail_handler.setLevel(logging.ERROR)
-        app.logger.addHandler(mail_handler)
-    if app.config['LOG_TO_STDOUT']:
-        stream_handler = logging.StreamHandler()
-        stream_handler.setLevel(logging.INFO)
-        app.logger.addHandler(stream_handler)
-    else:
-        if not os.path.exists('logs'):
-            os.mkdir('logs')
-        file_handler = RotatingFileHandler(
-            'logs/taskmate.log',
-            maxBytes=10240,
-            backupCount=10)
-        file_handler.setFormatter(logging.Formatter(
-            '%(asctime)s %(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'))
-        file_handler.setLevel(logging.INFO)
-        app.logger.addHandler(file_handler)
-        app.logger.setLevel(logging.INFO)
-        app.logger.info('Taskmate Startup')
+    from app.auth import bp as auth_bp
+    app.register_blueprint(auth_bp, url_prefix='/auth')
 
-from app import routes, models, errors
+    from app.user import bp as user_bp
+    app.register_blueprint(user_bp, url_prefix='/user')
+
+    from app.main import bp as main_bp
+    app.register_blueprint(main_bp)
+
+    app.config.from_object(Config)
+
+    db.init_app(app)
+    migrate.init_app(app, db, render_as_batch=True)
+    moment.init_app(app)
+    login.init_app(app)
+    mail.init_app(app)
+    boostrap.init_app(app)
+
+    if not app.debug:
+        if app.config['MAIL_SERVER']:
+            auth = None
+            if app.config['MAIL_USERNAME'] or app.config['MAIL_PASSWORD']:
+                auth = (app.config['MAIL_USERNAME'], app.config['MAIL_PASSWORD'])
+            secure = None
+            if app.config['MAIL_USE_TLS']:
+                secure = ()
+            mail_handler = SMTPHandler(
+                mailhost=(app.config['MAIL_SERVER'], app.config['MAIL_PORT']),
+                # fromaddr='noreply@' + app.config['MAIL_SERVER'],
+                fromaddr=app.config['MAIL_DEFAULT_SENDER'],
+                toaddrs=app.config['ADMINS'],
+                subject='Taskmate Failure',
+                credentials=auth, secure=secure)
+            mail_handler.setLevel(logging.ERROR)
+            app.logger.addHandler(mail_handler)
+        if app.config['LOG_TO_STDOUT']:
+            stream_handler = logging.StreamHandler()
+            stream_handler.setLevel(logging.INFO)
+            app.logger.addHandler(stream_handler)
+        else:
+            if not os.path.exists('logs'):
+                os.mkdir('logs')
+            file_handler = RotatingFileHandler(
+                'logs/taskmate.log',
+                maxBytes=10240,
+                backupCount=10)
+            file_handler.setFormatter(logging.Formatter(
+                '%(asctime)s %(levelname)s: %(message)s [in %(pathname)s:%(lineno)d]'))
+            file_handler.setLevel(logging.INFO)
+            app.logger.addHandler(file_handler)
+            app.logger.setLevel(logging.INFO)
+            app.logger.info('Taskmate Startup')
+    return app
+
+from app import models
