+++ app/auth/forms.py
@@ -1,10 +1,8 @@
 from flask_wtf import FlaskForm
-from wtforms import StringField, SubmitField, PasswordField, BooleanField,\
-    TextAreaField, DateField
+from wtforms import StringField, SubmitField, PasswordField, BooleanField
 from wtforms.validators import DataRequired, Length, EqualTo, Email, \
     Regexp, ValidationError
 from app.models import User
-from flask_wtf.file import FileField, FileAllowed, FileRequired
 
 
 # -----------------------
@@ -120,51 +118,4 @@ class VerifyForm(FlaskForm):
 
 # -----------------------
 # End of user registration
-# -----------------------
-
-
-# -----------------------
-# Create task
-# -----------------------
-
-
-class CreateTaskForm(FlaskForm):
-    """User can create a task"""
-    title = StringField(
-        'Title',
-        validators=[DataRequired(), Length(min=2, max=64)],
-        render_kw={'autofocus':True, 'placeholder': 'TaskMate Title'})
-    body = TextAreaField(
-        'Body',
-        validators=[DataRequired()],
-        render_kw={'autofocus':True, 'placeholder': 'Markdown enabled'})
-    due_date = DateField(
-        'Due Date', validators=[DataRequired()])
-    file = FileField(
-        'Attach File',
-        validators=[FileAllowed(['png', 'gif', 'jpg', 'jpeg', 'pdf'], 'PDF and pictures only!')])
-    submit = SubmitField('Create')
-
-
-
-class TaskCommentForm(FlaskForm):
-    """User can add a comment to a task"""
-    body = TextAreaField(
-        'Body',
-        validators=[DataRequired()],
-        render_kw={'autofocus':True, 'placeholder': 'Markdown enabled'})
-    submit_comment = SubmitField('Create')
-
-# -----------------------
-# End of create task
-# -----------------------
-
-
-# -----------------------
-# Edit profile
-# -----------------------
-
-
-# -----------------------
-# End of edit profile
 # -----------------------
