+++ app/errors/handlers.py
@@ -1,13 +1,14 @@
-from app import app, db
+from app import db
 from flask import render_template
+from app.errors import bp
 
 
-@app.errorhandler(404)
+@bp.errorhandler(404)
 def not_found(error):
     return render_template('errors/404.html', title='Page Not Found'), 404
 
 
-@app.errorhandler(500)
+@bp.errorhandler(500)
 def unexpected_error(error):
     db.session.rollback()
     return render_template('errors/500.html', title='Unexpected Error'), 500
