+++ app/models.py
@@ -1,10 +1,11 @@
-from app import db, login, app
+from app import db, login
 from flask_login import UserMixin
 from werkzeug.security import generate_password_hash, check_password_hash
 from datetime import datetime
 import jwt
 from time import time
 from hashlib import md5
+from flask import current_app
 
 
 @login.user_loader
@@ -34,14 +35,14 @@ class User(db.Model, UserMixin):
     def get_reset_password_token(self, expires_in=600):
         return jwt.encode(
             {'reset_password': self.id, 'exp': time() + expires_in},
-            app.config['SECRET_KEY'], algorithm='HS256')
+            current_app.config['SECRET_KEY'], algorithm='HS256')
 
     @staticmethod
     def verify_reset_password_token(token):
         try:
             id = jwt.decode(
                 token,
-                app.config['SECRET_KEY'],
+                current_app.config['SECRET_KEY'],
                 algorithms=['HS256'])['reset_password']
         except:
             return
