+++ app/templates/auth/login.html
@@ -17,7 +17,7 @@
       <br>
       <small>
         Forgot password? 
-        <a href="{{ url_for('request_password_reset') }}">Reset here</a>.
+        <a href="{{ url_for('auth.request_password_reset') }}">Reset here</a>.
       </small>
     </div>   
   </div>
