--- app/templates/authenticated_user/edit_task.html
@@ -1,40 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-<div class="vh-100 d-flex align-items-center justify-content-center">
-  <div class="w-75">
-    <div class="card text-bg-secondary">
-      <div class="card-header">
-        <h3 class="card-title mb-3 font-weight-normal text-center">{% block title %}Edit Task 📝{% endblock %}
-        </h3>
-      </div>
-      <div class="card-body">
-        <p class="font-weight-normal text-center">The more detail the better!</p>  
-        <hr>
-        <form method="post">
-          <div class="row">
-            <div class="col-md-6 mb-3">
-              <label for="title">Title</label>
-              <input name="title" id="title" class="form-control" value="{{ task['title'] }}" required>
-            </div>
-            <div class="col-md-6 mb-3">
-              <label for="due_date">Due Date</label>
-              <input type="date" class="form-control" id="due_date" name="due_date" value="{{ task['due_date'] }}">
-            </div>
-          </div>
-          <div class="mb-3">
-            <label for="body">Description</label>
-            {% if task['body'] is not none %}   
-            <textarea name="body" class="form-control" id="body">{{ task['body'] }}</textarea>
-            {% else %}
-            <textarea name="body" class="form-control" id="body"></textarea>
-            {% endif %}
-          </div>
-
-          <input type="submit" value="Save">
-        </form>
-      </div>
-    </div>
-  </div>
-</div>
-{% endblock %}
