+++ app/templates/authenticated_user/home.html
@@ -46,7 +46,7 @@
                 <div class="card">
                     <div class="card-body">
                     <h3 class="card-title">
-                        <a href="{{ url_for('task_details', title=task.title) }}">{{ task.title }}</a>
+                        <a href="{{ url_for('user.task_details', title=task.title) }}">{{ task.title }}</a>
                     </h3>
                     <h6 class="card-subtitle mb-2 text-muted">Due Date: <strong>{{ moment(task.due_date).format('ll') }}</strong> - 
                     {% if task.status == 'Active' %}
@@ -71,7 +71,7 @@
                 <div class="card">
                     <div class="card-body">
                     <h3 class="card-title">
-                        <a href="{{ url_for('task_details', title=task.title) }}">{{ task.title }}</a>
+                        <a href="{{ url_for('user.task_details', title=task.title) }}">{{ task.title }}</a>
                     </h3>
                     <h6 class="card-subtitle mb-2 text-muted">Due Date: <strong>{{ moment(task.due_date).format('ll') }}</strong> - <span class="label label-danger">{{ task.status }}</span></h6>
                     <p class="card-text">{{ task.body }}</p>
