+++ app/templates/base.html
@@ -26,20 +26,20 @@
                   <span class="icon-bar"></span>
                   <span class="icon-bar"></span>
               </button>
-              <a class="navbar-brand" href=" {{ url_for('index') }} ">TaskMate 📝</a>
+              <a class="navbar-brand" href=" {{ url_for('main.index') }} ">TaskMate 📝</a>
           </div>
           <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">                          
                 {% if current_user.is_authenticated %}
                   <ul class="nav navbar-nav">
                     <li class="nav-item">                    
-                      <a class="nav-link" href="{{ url_for('home') }}">
+                      <a class="nav-link" href="{{ url_for('user.home') }}">
                         <button type="button" class="btn text-white bg-primary">
                           Task Details
                         </button>
                       </a>
                     </li>
                     <li class="nav-item">
-                      <a href="{{ url_for('create_task') }}">
+                      <a href="{{ url_for('user.create_task') }}">
                         <button type="button" class="btn text-white bg-success">
                           Create a task
                         </button>
@@ -65,7 +65,7 @@
                       </span>
                     </li>                  
                     <li class="nav-item">
-                      <a href="{{ url_for('logout') }}">
+                      <a href="{{ url_for('auth.logout') }}">
                         <button type="button" class="btn text-white bg-warning">
                           Logout
                         </button>
@@ -74,18 +74,26 @@
                 {% else %}
                   <ul class="nav navbar-nav navbar-right">
                     <li class="nav-item">
-                      <a class="btn btn-outline-primary mr-1" href="{{ url_for('register') }}">
-                        <img width="25" height="25" src="https://img.icons8.com/sf-regular-filled/48/add-user-male.png"
-                          alt="add-user-male" /> Sign up
+                      <a href="{{ url_for('auth.register') }}">
+                        <button type="button" class="btn btn-outline-info">
+                          <img width="25" height="25" src="https://img.icons8.com/sf-regular-filled/48/add-user-male.png" alt="add-user-male" /> Sign up
+                        </button>                        
                       </a>
                     </li>                
                     <li class="nav-item">
-                      <a class="btn btn-outline-primary" href="{{ url_for('login') }}">
-                        <img width="25" height="25" src="https://img.icons8.com/sf-regular-filled/23/login-rounded-down.png"
-                          alt="login-rounded-down" /> Login
+                      <a href="{{ url_for('auth.login') }}">
+                        <button type="button" class="btn btn-outline-warning">
+                          <img width="25" height="25" src="https://img.icons8.com/sf-regular-filled/23/login-rounded-down.png" alt="login-rounded-down" /> Login
+                        </button>                        
                       </a>
                     </li>
-                    <li class="nav-item"><a href=" https://www.github.com/michaelzdrav/taskmate " target="_blank">Source Code</a></li>
+                    <li class="nav-item">
+                      <a href=" https://www.github.com/michaelzdrav/taskmate " target="_blank">
+                        <button type="button" class="btn btn-outline-warning">
+                          <img width="25" height="25" src="https://img.icons8.com/?size=512&id=62856&format=png" alt="login-rounded-down" /> Source Code
+                        </button>     
+                      </a>
+                    </li>
                   </ul>
                 {% endif %}                    
           </div>
