+++ app/templates/emails/auth/thank_you_signup.html
@@ -2,7 +2,7 @@
 <p>
     Thank you for signing up for TaskMate, your free task management tool. Login to continue 
     by clicking 
-    <a href="{{ url_for('login', _external=True) }}">here</a>
+    <a href="{{ url_for('auth.login', _external=True) }}">here</a>
 </p>
 <p>Alternatively, check for the login button in the navigation bar of the app.</p>
 <p>Sincerely,</p>
