+++ app/templates/emails/auth/thank_you_signup.txt
@@ -3,7 +3,7 @@ Dear {{ username }},
 Thank you for signing up for TaskMate, your free task management tool. Login to continue 
 by clicking 
 
-{{ url_for('login', _external=True) }}
+{{ url_for('auth.login', _external=True) }}
 
 Alternatively, check for the login button in the navigation bar of the app.
