--- app/templates/emails/new_task.html
@@ -1,36 +1,0 @@
-<head>
-    <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/css/bootstrap.min.css">
-</head>
-
-<body>
-    <p>Dear {{ user.username }},</p>
-    <p>Here are the details of a new task you have created.</p>
-    </p>
-    <div class="container">
-        <table class="table">
-            <thead>
-                <tr>
-                    <th>Title:</th>
-                    <td>{{ body.title }}</td>
-                </tr>
-            </thead>
-            <tbody>
-                <tr>
-                    <th>Due Date:</th>
-                    <td>{{ body.due_date }}</td>
-                </tr>
-                <tr>
-                    <th>Description:</th>
-                    <td>{{ body.description }}</td>
-                </tr>
-            </tbody>
-        </table>
-    </div>
-    <p>Sincerely,</p>
-
-    </hr>
-    <p>The TaskMate Team.</p>
-
-    <!-- Bootstrap JS -->
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.1/dist/js/bootstrap.bundle.min.js"></script>
-</body>
