+++ app/templates/emails/user/overdue_task.html
@@ -9,7 +9,7 @@
 </p>
 <p>
     You can review this task and take action by clicking 
-    <a href="{{ url_for('home', _external=True) }}">here</a>.
+    <a href="{{ url_for('user.home', _external=True) }}">here</a>.
 </p>
 <p>Sincerely,</p>
 <p>The TaskMate Team.</p>
