+++ app/templates/errors/404.html
@@ -11,7 +11,7 @@
 <div class="row text-center">
   <div class="col-md-4"></div>
   <div class="col-md-4">
-    Return to the <a href="{{ url_for('index') }}">home page</a>
+    Return to the <a href="{{ url_for('main.index') }}">home page</a>
   </div>
   <div class="col-md-4"></div>
 </div>
@@ -29,7 +29,7 @@
 <div class="row text-center">
   <div class="col-md-4"></div>
   <div class="col-md-4">
-    Return to the <a href="{{ url_for('index') }}">home page</a>
+    Return to the <a href="{{ url_for('main.index') }}">home page</a>
   </div>
   <div class="col-md-4"></div>
 </div>
