+++ app/user/__init__.py
@@ -1,0 +1,7 @@
+from flask import Blueprint
+
+
+bp = Blueprint('user', __name__)
+
+
+from app.user import routes
