+++ main.py
@@ -1,6 +1,7 @@
-from app import app, db, cli
+from app import create_app, db, cli
 from app.models import User, Task, TaskComment
 
+app = create_app()
 cli.register(app)
 
 @app.shell_context_processor
